Belafonte catches otherwise uncaught exceptions
If an exception is raised within #setup, #handle, or a subcommand,
but is not caught in any of those places, it is caught by
Belafonte itself.

In the end, it's my hope that this will never actually be leaned
upon, but doing this is way better, typically, than showing an
ugly damned stack trace.